Code/Backend/requirements.txt
# DagBot Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pyyaml==6.0.2
aiosqlite==0.20.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
pydantic>=2.10.0
httpx>=0.27.0
